Organize common components into directories, udpate imports
