+++ src/react/components/common/condensedList/condensedList.scss
