+++ src/react/components/common/condensedList/condensedList.tsx
