+++ src/react/components/common/connectionPicker/connectionPicker.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IConnection } from "../../../models/applicationState";
+import { IConnection } from "../../../../models/applicationState";
 import { Link } from "react-router-dom";
 
 interface IConnectionPickerProps {
