+++ src/react/components/common/customField/customField.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { FieldProps } from "react-jsonschema-form";
-import Guard from "../../../common/guard";
+import Guard from "../../../../common/guard";
 
 export default function CustomField(Widget: any, mapProps?: (props: FieldProps) => any) {
     Guard.null(Widget);
