+++ src/react/components/common/localFolderPicker/localFolderPicker.test.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import LocalFolderPicker from "./localFolderPicker";
 import { mount } from "enzyme";
 
-jest.mock("../../../providers/storage/localFileSystemProxy");
-import { LocalFileSystemProxy } from "../../../providers/storage/localFileSystemProxy";
+jest.mock("../../../../providers/storage/localFileSystemProxy");
+import { LocalFileSystemProxy } from "../../../../providers/storage/localFileSystemProxy";
 
 describe("Local Folder Picker Component", () => {
     const onChangeHandler = jest.fn();
